Setup the base UI
I still need to clean this all up, fix auto layout bugs and performance
analysis but below is summary of what changed and why.

o To make the UI simple use a navigation controller and set the right
item to a ui switch.
o Add a extension for now just using one color but plan on populating it
with more later.
o Handle image downloading, for now we are not caching anything. Ideally
we should persist for a short while to reduce hitting the network.
Memory usage looks okay so far.
o Wire up all of the UI interactions with stubs for now.
o Also add todo notes for later.

Taking a dinner break soon, but will get to the polish round in a couple
of hours ;)

Closes: #6 (Create initial draft of view controllers)