Split model for api and ui
Using structures was fine in the beginning but now that we are mutating
the favorite state, passing by reference makes it easier to grapple with
the code.

We are not marking the ads but nothing is persisted, everything
disappears between launches. So next we should store the view model and
images.

Closes: #8 (handle target for add button is pressed)