Use cache for the images
hide the downloading inside of the APIHandler, we are just interested in
the completion handler. Also added a detail view for showing more of the
advertisement. As noted in one of the comments, I was getting stuck with
autolayout topics, so decide to split it into two view controllers
instead. I have roughly 10 hours left. Hopefully I can get to the
storage parts, update README and polish overall.