store the model used by the UI for rendering
We are using a flat file on the system, which is working but should be
replaced with either db solution like SQLite[0], Realm[1], Sync[2], etc.
I am doing it this way to save some time for other topics.

Also added updated tests to check this and use new class names. Totally
missed the naming breakage earlier...

[0]: https://www.sqlite.org/index.html
[1]: https://github.com/realm/realm-cocoa
[2]: https://github.com/3lvis/Sync

Closes: #9 (persist minimal data required to present the something)