Add the favoriteButton to detail vc
While not ideal, copy pasting it over does work but moving back to the
parent the collection view does not redraw. The state is set, but I
don't want to trigger full redraw so will think about it some more and
come up with something. One idea was to use the indexpath but we would
need to identify the item that changed.