Improve the distinction between favorite ads and regular ones
Note this is mainly shuffling code around and the minimal UI change is just to
the default order. So instead of fetching ads immeditaely we check for the
favorites, if available only load them. Otherwise fetch remote ads. Non
favorite ads are no longer stored in the home directory.

Other misc changes are

* Make the save code private and only access it by designated methods.  Ideally
it should not save at every change but batch them up. Will look into that
later.
* Reword some of the comments
* Slim down the setup method in AdsCollectionViewController