Only rely on the datasource for initialzing the currentAd variable
Since it can be reset after our setup has complete we can't assume there
will be a valid ad on subsequent calls. This is due to the way the
removal is done, for more details read the following commit c29883b36087
(Use a delegate as well for toggling favorite state, 2018-11-21).