Remove all usage of .reloadData() on the collectionView
Apple documentation says to use it sparingly when reloading all items in
a collection view[0]. Now that we are handling insertion and deletion
it's really not needed anymore. Getting this to work did require some
reshuffling of the code and handling a couple of cases where there could
be mismatch between visible ads and the updates.

Also the ads are now sorted in order to prevent extra reload cycles and
stale cells.

[0]: https://developer.apple.com/documentation/uikit/uicollectionview/1618078-reloaddata