Rename EmptyFavoritesView to EmptyCollectionView
Making the title and body variable allows us to reuse them for more
states.  I was working on the error view but after meditating on it will
use this view instead and in follow up maybe add a "call for action"
button to retry the request.

While at it try to be consistent with the usage of message / body and
title / header.